passive-reconnaissance/requirements.txt
dnspython>=2.6.0
python-whois>=0.9.4
requests>=2.31.0
